requirements.txt
Flask
requests
dnspython
mitmproxy